+++ java/Thuchanh/BagQueueStack/Product.java
@@ -2,15 +2,15 @@
 
 public class Product implements Expression{
 
-	private final Expression leftExpression;
-	private final Expression rightExpression;
+    private final Expression leftExpression;
+    private final Expression rightExpression;
 
-	public Product(Expression leftExpression,Expression rightExpression ){
-		this.leftExpression = leftExpression;
-		this.rightExpression = rightExpression;
-	}
-	@Override
-	public int interpret() {
-		return leftExpression.interpret() * rightExpression.interpret();
-	}
+    public Product(Expression leftExpression,Expression rightExpression ){
+        this.leftExpression = leftExpression;
+        this.rightExpression = rightExpression;
+    }
+    @Override
+    public double interpret() {
+        return leftExpression.interpret() * rightExpression.interpret();
+    }
 }
